refactor: use future.js library
