+++ package.json
@@ -23,6 +23,7 @@
     "url": "https://github.com/ketch-sdk/ketch-tag/issues"
   },
   "devDependencies": {
+    "@ketch-com/future": "^1.0.1",
     "@ketch-sdk/ketch-types": "^0.3.0",
     "@ketch-sdk/ketch-web-api": "^1.0.5",
     "@types/fetch-mock": "^7.3.5",
@@ -36,6 +37,7 @@
     "eslint-plugin-import": "^2.26.0",
     "eslint-plugin-jest": "^27.1.3",
     "eslint-plugin-prettier": "^4.2.1",
+    "events": "^3.3.0",
     "jest": "^29.2.1",
     "jest-environment-jsdom": "^29.2.1",
     "jest-fetch-mock": "^3.0.3",
